// AdminPortal/src/app/model/mail-config.model.ts
export class MailConfig {
    UserID: number;
    MailBoxName: string;
    SMTPServer: string;
    SMTPPort: number;
    UserName: string;
    Password: string;
    IsSSL: boolean;
    MailFrom: string;
    ReplyTo: string;
    IsActive: boolean;
    memRefNo:string;
    DisplayName:string;
    Company:string;
}